pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "butter-ai"
version = "0.1.0"
description = "Butter AI Games"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0.1",
    "torch==2.1.2",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
gpu = [
    "torch==2.1.2+cu118",
]

[tool.setuptools]
package-dir = {"" = "projects"}
packages = ["common", "love-letter"]

[tool.pytest.ini_options]
pythonpath = [
    "projects"
]
